+++ package.json
@@ -32,10 +32,15 @@
   "dependencies": {
     "@babel/core": "^7.12.9",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@vue/compiler-sfc": "^3.0.3",
     "babel": "^6.23.0",
     "babel-loader": "^8.2.2",
     "css-loader": "^5.0.1",
     "style-loader": "^2.0.0",
+    "vue": "^2.6.12",
+    "vue-loader": "^15.9.5",
+    "vue-style-loader": "^4.1.2",
+    "vue-template-compiler": "^2.6.12",
     "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^5.4.0"
   }
