+++ public/index.html
@@ -10,35 +10,8 @@
 </head>
 
 <body>
-    <header>
-        <h2>МАГАЗИН</h2>
-    </header>
-
     <main>
-        <br>
-        <p><b>ТОВАРЫ МАГАЗИНА</b></p>
-        <br>
-        <div class="productsList"></div>
-
-        <br>
-        <button type="button" name="btnAddNextChunk">Добавить ещё</button>
-        <br>
-
-        <br>
-        <p><b>ТОВАРЫ В КОРЗИНЕ</b></p>
-        <br>
-        <div class="basketList"></div>
-
-        <br>
-        <p><b>КОРЗИНА</b></p>
-        <br>
-        <div class="basket"></div>
-
-        <br>
-        <button name="ClearBasket">Очистить корзину</button>
-
-        <br>
-        <br>
+        
         <a href="./Forms/feedback.html">Обратная связь</a>
     </main>
