+++ public/scripts/shoppack.js.LICENSE.txt
@@ -1,0 +1,5 @@
+/*!
+ * Vue.js v2.6.12
+ * (c) 2014-2020 Evan You
+ * Released under the MIT License.
+ */
