+++ src/scripts/Item.vue
@@ -1,0 +1,17 @@
+<template>
+    <div>{{ id }}. {{ name }} - {{ price }}</div>
+</template>
+
+<script>
+export default {
+    props: {
+        id: Number,
+        name: String,
+        price: Number
+    }
+}
+</script>
+
+<style >
+
+</style>
