+++ webpack.config.js
@@ -1,10 +1,16 @@
 const path = require("path")
+const VueLoaderPlugin = require('vue-loader/lib/plugin')
 
 module.exports = {
     entry: "./src/scripts/Shop.js",
     output: {
         path: path.resolve(__dirname, "./public/scripts"),
         filename: "shoppack.js"
+    },
+    resolve: {
+        alias: {
+            vue: 'vue/dist/vue.esm.js'
+        }
     },
     module: {
         rules: [
@@ -18,9 +24,23 @@ module.exports = {
                 test: /\.css$/,
                 use: [
                     { loader: "style-loader" },
-                    { loader: "css-loader" }
+                    { 
+                        loader: "css-loader",
+                        options: {
+                            modules: true
+                        }
+                    }
+                ]
+            },
+            {
+                test: /\.vue$/,
+                use: [
+                    { loader: "vue-loader" }
                 ]
             }
         ]
-    }
+    },
+    plugins: [
+        new VueLoaderPlugin()
+    ]
 }
